refactor: Own module for text rendering.
